rename: verifyAccessToken.js를 authorizationMiddleware.js로 변경
